Allow to not depend on offscreen_gl_context.
This PR allows (via feature-gate) to compile out enough WebGL-related code in
order to not depend in offscreen_gl_context.

Fixes bug 1331546
Fixes https://github.com/emilio/rust-offscreen-rendering-context/issues/77
