+++ .travis.yml
@@ -22,6 +22,7 @@ before_install:
 script:
   - (cd webrender_traits && cargo test --verbose --features "ipc")
   - (cd webrender_traits && cargo test --verbose)
+  - (cd webrender && cargo build --verbose --no-default-features --features serde_derive)
   - (cd webrender && cargo test --verbose)
   - (cd sample && cargo test --verbose)
   - (cd replay && cargo test --verbose)
