+++ webrender/Cargo.toml
@@ -8,10 +8,11 @@ build = "build.rs"
 workspace = ".."
 
 [features]
-default = ["codegen", "freetype-lib"]
+default = ["codegen", "freetype-lib", "webgl"]
 codegen = ["webrender_traits/codegen"]
 freetype-lib = ["freetype/servo-freetype-sys"]
 serde_derive = ["webrender_traits/serde_derive"]
+webgl = ["offscreen_gl_context", "webrender_traits/webgl"]
 
 [dependencies]
 app_units = "0.3"
@@ -24,7 +25,7 @@ gleam = "0.2.30"
 lazy_static = "0.2"
 log = "0.3"
 num-traits = "0.1.32"
-offscreen_gl_context = {version = "0.5", features = ["serde_serialization", "osmesa"]}
+offscreen_gl_context = {version = "0.5", features = ["serde_serialization", "osmesa"], optional = true}
 time = "0.1"
 threadpool = "1.3.2"
 webrender_traits = {path = "../webrender_traits", default-features = false}
