+++ webrender/src/render_backend.rs
@@ -4,8 +4,7 @@
 
 use byteorder::{LittleEndian, ReadBytesExt};
 use frame::Frame;
-use internal_types::{FontTemplate, GLContextHandleWrapper, GLContextWrapper};
-use internal_types::{SourceTexture, ResultMsg, RendererFrame};
+use internal_types::{FontTemplate, SourceTexture, ResultMsg, RendererFrame};
 use profiler::BackendProfileCounters;
 use record;
 use resource_cache::ResourceCache;
@@ -16,11 +15,13 @@ use std::io::{Cursor, Read};
 use std::sync::{Arc, Mutex};
 use std::sync::mpsc::Sender;
 use texture_cache::TextureCache;
+use webgl_types::{GLContextHandleWrapper, GLContextWrapper};
 use webrender_traits::{ApiMsg, AuxiliaryLists, BuiltDisplayList, IdNamespace, ImageData};
 use webrender_traits::{RenderNotifier, RenderDispatcher, WebGLCommand, WebGLContextId};
 use webrender_traits::channel::{PayloadHelperMethods, PayloadReceiver, PayloadSender, MsgReceiver};
 use webrender_traits::{VRCompositorCommand, VRCompositorHandler};
 use tiling::FrameBuilderConfig;
+#[cfg(feature = "webgl")]
 use offscreen_gl_context::GLContextDispatcher;
 
 /// The render backend is responsible for transforming high level display lists into
@@ -42,12 +43,16 @@ pub struct RenderBackend {
     frame: Frame,
 
     notifier: Arc<Mutex<Option<Box<RenderNotifier>>>>,
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     webrender_context_handle: Option<GLContextHandleWrapper>,
     webgl_contexts: HashMap<WebGLContextId, GLContextWrapper>,
     current_bound_webgl_context_id: Option<WebGLContextId>,
     enable_recording: bool,
+
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     main_thread_dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>,
 
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     next_webgl_id: usize,
 
     vr_compositor_handler: Arc<Mutex<Option<Box<VRCompositorHandler>>>>
@@ -254,6 +259,7 @@ impl RenderBackend {
                             tx.send(self.frame.get_scroll_layer_state())
                               .unwrap()
                         }
+                        #[cfg(feature = "webgl")]
                         ApiMsg::RequestWebGLContext(size, attributes, tx) => {
                             if let Some(ref wrapper) = self.webrender_context_handle {
                                 let dispatcher: Option<Box<GLContextDispatcher>> = if cfg!(target_os = "windows") {
@@ -289,6 +295,7 @@ impl RenderBackend {
                                 tx.send(Err("Not implemented yet".to_owned())).unwrap();
                             }
                         }
+                        #[cfg(feature = "webgl")]
                         ApiMsg::ResizeWebGLContext(context_id, size) => {
                             let ctx = self.webgl_contexts.get_mut(&context_id).unwrap();
                             ctx.make_current();
@@ -480,10 +487,12 @@ impl RenderBackend {
     }
 }
 
+#[cfg(feature = "webgl")]
 struct WebRenderGLDispatcher {
     dispatcher: Arc<Mutex<Option<Box<RenderDispatcher>>>>
 }
 
+#[cfg(feature = "webgl")]
 impl GLContextDispatcher for WebRenderGLDispatcher {
     fn dispatch(&self, f: Box<Fn() + Send>) {
         let mut dispatcher = self.dispatcher.lock();
