+++ webrender/src/resource_cache.rs
@@ -186,6 +186,7 @@ pub struct ResourceCache {
     cached_images: ResourceClassCache<ImageRequest, CachedImageInfo>,
 
     // TODO(pcwalton): Figure out the lifecycle of these.
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     webgl_textures: HashMap<WebGLContextId, WebGLTexture, BuildHasherDefault<FnvHasher>>,
 
     font_templates: HashMap<FontKey, FontTemplate, BuildHasherDefault<FnvHasher>>,
@@ -298,6 +299,7 @@ impl ResourceCache {
         println!("Delete the non-exist key:{:?}", image_key);
     }
 
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     pub fn add_webgl_texture(&mut self, id: WebGLContextId, texture_id: SourceTexture, size: DeviceIntSize) {
         self.webgl_textures.insert(id, WebGLTexture {
             id: texture_id,
@@ -305,6 +307,7 @@ impl ResourceCache {
         });
     }
 
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     pub fn update_webgl_texture(&mut self, id: WebGLContextId, texture_id: SourceTexture, size: DeviceIntSize) {
         let webgl_texture = self.webgl_textures.get_mut(&id).unwrap();
 
@@ -449,6 +452,7 @@ impl ResourceCache {
     }
 
     #[inline]
+    #[cfg_attr(not(feature = "webgl"), allow(dead_code))]
     pub fn get_webgl_texture(&self, context_id: &WebGLContextId) -> CacheItem {
         let webgl_texture = &self.webgl_textures[context_id];
         CacheItem {
