+++ webrender_traits/Cargo.toml
@@ -8,10 +8,11 @@ build = "build.rs"
 workspace = ".."
 
 [features]
-default = ["codegen"]
+default = ["codegen", "webgl"]
 nightly = ["euclid/unstable", "serde/unstable"]
 codegen = ["serde_codegen", "serde_codegen/with-syntex"]
 ipc = ["ipc-channel"]
+webgl = ["offscreen_gl_context"]
 
 [dependencies]
 app_units = "0.3.0"
@@ -19,7 +20,7 @@ byteorder = "0.5"
 euclid = "0.10.3"
 gleam = "0.2"
 heapsize = "0.3.6"
-offscreen_gl_context = {version = "0.5.0", features = ["serde_serialization"]}
+offscreen_gl_context = {version = "0.5.0", features = ["serde_serialization"], optional = true}
 serde = "0.8"
 serde_derive = {version = "0.8", optional = true}
 ipc-channel = { version = "0.5.0", optional = true }
